LayoutTests/css3/flexbox/css-properties-expected.txt

Tests being able to set the display to -webkit-flexbox and -webkit-inline-flexbox.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS flexbox.style.display is "-webkit-flexbox"
PASS flexbox.style.display is "inline"
PASS flexbox.style.display is "-webkit-inline-flexbox"
PASS flexbox.style.display is "-webkit-inline-flexbox"
PASS flexbox.style.display is "block"
PASS flexitem.style.webkitFlexOrder is ""
PASS flexitem.style.webkitFlexOrder is "2"
PASS flexitem.style.webkitFlexOrder is "-1"
PASS flexitem.style.webkitFlexOrder is "0"
PASS flexitem.style.webkitFlexOrder is "0"
PASS flexitem.style.webkitFlexOrder is "0"
PASS flexitem.style.webkitFlexOrder is ""
PASS flexbox.style.webkitFlexPack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "start"
PASS flexbox.style.webkitFlexPack is ""
PASS flexbox.style.webkitFlexPack is "start"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "start"
PASS flexbox.style.webkitFlexPack is "end"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "end"
PASS flexbox.style.webkitFlexPack is "center"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "center"
PASS flexbox.style.webkitFlexPack is "justify"
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "justify"
PASS flexbox.style.webkitFlexPack is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexPack is "start"
PASS flexbox.style.webkitFlexAlign is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "stretch"
PASS flexbox.style.webkitFlexAlign is ""
PASS flexbox.style.webkitFlexAlign is "before"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "before"
PASS flexbox.style.webkitFlexAlign is "after"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "after"
PASS flexbox.style.webkitFlexAlign is "middle"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "middle"
PASS flexbox.style.webkitFlexAlign is "stretch"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "stretch"
PASS flexbox.style.webkitFlexAlign is "baseline"
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "baseline"
PASS flexbox.style.webkitFlexAlign is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexAlign is "stretch"
PASS flexbox.style.webkitFlexFlow is ""
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row"
PASS flexbox.style.webkitFlexFlow is ""
PASS flexbox.style.webkitFlexFlow is "row"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row"
PASS flexbox.style.webkitFlexFlow is "row-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "row-reverse"
PASS flexbox.style.webkitFlexFlow is "column"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column"
PASS flexbox.style.webkitFlexFlow is "column-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "column-reverse"
PASS flexbox.style.webkitFlexFlow is "horizontal"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "horizontal"
PASS flexbox.style.webkitFlexFlow is "horizontal-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "horizontal-reverse"
PASS flexbox.style.webkitFlexFlow is "horizontal-ltr"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "horizontal-ltr"
PASS flexbox.style.webkitFlexFlow is "horizontal-rtl"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "horizontal-rtl"
PASS flexbox.style.webkitFlexFlow is "vertical"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "vertical"
PASS flexbox.style.webkitFlexFlow is "vertical-reverse"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "vertical-reverse"
PASS flexbox.style.webkitFlexFlow is "vertical-ttb"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "vertical-ttb"
PASS flexbox.style.webkitFlexFlow is "vertical-btt"
PASS window.getComputedStyle(flexbox, null).webkitFlexFlow is "vertical-btt"
PASS successfullyParsed is true

TEST COMPLETE
